refactor(naming): 파일·폴더명 통일 및 데드코드 제거
- Excutives → Executives (페이지·리소스 폴더)
- components/apply → components/Apply
- resources/Executive → resources/Executives
- EmailButton: 함수명 CopyEmail → EmailButton 통일
- LectureDetail: 컴포넌트명 LectureDetailCombined → LectureDetail
- MemberCard: 미사용 CopyEmail import 제거
- ProtectedRoute 미사용으로 삭제
