+++ getit-web/src/pages/Public/Apply.jsx
@@ -5,9 +5,9 @@ import { useAppStore } from '../../hooks/appStore';
 import { useAuth } from '../../hooks/useAuth';
 import { MESSAGES, APPLY_ANNOUNCE_DATE } from '../../constants';
 import { answersToPayload, payloadToAnswers, isApplicationComplete } from '../../utils/applyForm';
-import ApplyHeader from '../../components/apply/ApplyHeader';
-import QuestionField from '../../components/apply/QuestionField';
-import SubmitButton from '../../components/apply/SubmitButton';
+import ApplyHeader from '../../components/Apply/ApplyHeader';
+import QuestionField from '../../components/Apply/QuestionField';
+import SubmitButton from '../../components/Apply/SubmitButton';
 import questionData from '../../resources/Apply/question.json';
 
 const Apply = () => {
